Exclusive access to the storage (lock)
A container can't can used by more than one NBD client at the same
time. For that purpose a 'client' flag is set in the container
metadata to "lock" the storage.

Some extra work may be needed to ensure that the storage is unlocked
in case the server crashes. swiftnbd-setup with -f flag can be used
to unlock a storage.